# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Nodejs-Case-Study-Sagnik",
        "description": "User Auth & Roles relation",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [{
        "url": "/",
        "description": "Local Dev"
    }],
    "tags": [{
        "name": "Users",
        "description": "API for users in the system"
    }],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/api/test/user": {
            "get": {
                "tags": [
                    "USER"
                ],
                "summary": "Get User Info by Id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
        "/api/test/pm": {
            "get": {
                "tags": [
                    "PM"
                ],
                "summary": "Get PM Info by ID | Restricted Access",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
        "/api/test/admin": {
            "get": {
                "tags": [
                    "ADMIN"
                ],
                "summary": "Get ADMIN Info by ID | Restricted Access",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
        "/api/auth/signin": {
            "post": {
                "tags": [
                    "USER"
                ],
                "summary": "Authorize an user",
                "requestBody": {
                    "description": "User Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Auth"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Wrong Password."
                    }
                }
            }
        },
        "/api/auth/signup": {
            "post": {
                "tags": [
                    "USER"
                ],
                "summary": "Create a new user",
                "requestBody": {
                    "description": "User Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Signup"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Username/Email exists"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "User": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "user": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        },
                        "username": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "roles": {
                            "type": "array",
                            "properties": {
                                "id": {
                                    "type": "number"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "user_roles": {
                                    "type": "array",
                                    "properties": {
                                        "userId": "number",
                                        "roleId": "number"
                                    }
                                }
                            }
                        }
                    }
                }

            }
        },
        "Auth": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "Signup": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 3,
                    "items": {
                        "type": "string"
                    },
                    "uniqueItems": true
                }
            }
        }
    }
}